+++ .dotnet/src/Custom/Chat/ChatOutputPrediction.cs
@@ -1,0 +1,19 @@
+using System;
+using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
+
+namespace OpenAI.Chat;
+
+[CodeGenModel("ChatOutputPrediction")]
+public partial class ChatOutputPrediction
+{
+    // CUSTOM: Rename; add public visibility for derived type discriminator insight.
+    [CodeGenMember("Type")]
+    public ChatOutputPredictionKind Kind { get; }
+
+    public static ChatOutputPrediction CreateStaticContentPrediction(IEnumerable<ChatMessageContentPart> contentParts)
+        => new InternalChatOutputPredictionContent(new ChatMessageContent(contentParts));
+
+    public static ChatOutputPrediction CreateStaticContentPrediction(string content)
+        => new InternalChatOutputPredictionContent([ChatMessageContentPart.CreateTextPart(content)]);
+}