+++ .dotnet/src/Custom/Chat/ChatOutputPredictionKind.cs
@@ -1,0 +1,13 @@
+using System;
+using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
+
+namespace OpenAI.Chat;
+
+[CodeGenModel("ChatOutputPredictionType")]
+public readonly partial struct ChatOutputPredictionKind
+{
+    // CUSTOM: Rename for clarity.
+    [CodeGenMember("Content")]
+    public static ChatOutputPredictionKind StaticContent { get; } = new ChatOutputPredictionKind(ContentValue);
+}